chore: rename recoveyMode package
